requirement.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[argon2]
python-jose[crptography]
python-multipart
pydantic[email]